remove fhirTaskPlanWorker and move logic in CarePlanGenerator
