+++ android/engine/src/main/java/org/smartregister/fhircore/engine/task/FhirCarePlanGenerator.kt
@@ -43,9 +43,12 @@ import org.smartregister.fhircore.engine.data.local.DefaultRepository
 import org.smartregister.fhircore.engine.util.extension.asReference
 import org.smartregister.fhircore.engine.util.extension.encodeResourceToString
 import org.smartregister.fhircore.engine.util.extension.extractId
+import org.smartregister.fhircore.engine.util.extension.hasPastEnd
+import org.smartregister.fhircore.engine.util.extension.hasStarted
 import org.smartregister.fhircore.engine.util.extension.isIn
 import org.smartregister.fhircore.engine.util.extension.referenceValue
 import org.smartregister.fhircore.engine.util.extension.setPropertySafely
+import org.smartregister.fhircore.engine.util.extension.toCoding
 import org.smartregister.fhircore.engine.util.helper.TransformSupportServices
 import timber.log.Timber
 
@@ -168,6 +171,8 @@ constructor(
 
     if (carePlanModified) saveCarePlan(output)
 
+    updateTaskStatus()
+
     return if (output.hasActivity()) output else null
   }
 
@@ -218,4 +223,47 @@ constructor(
 
   suspend fun getTask(id: String) =
     kotlin.runCatching { fhirEngine.get<Task>(id) }.onFailure { Timber.e(it) }.getOrNull()
+
+  suspend fun updateTaskStatus() {
+    Timber.i("Starting task scheduler")
+
+    // TODO also filter by date range for better performance
+    fhirEngine
+      .search<Task> {
+        filter(
+          Task.STATUS,
+          { value = of(TaskStatus.REQUESTED.toCoding()) },
+          { value = of(TaskStatus.READY.toCoding()) },
+          { value = of(TaskStatus.ACCEPTED.toCoding()) },
+          { value = of(TaskStatus.INPROGRESS.toCoding()) },
+          { value = of(TaskStatus.RECEIVED.toCoding()) },
+        )
+      }
+      .forEach { task ->
+        if (task.hasPastEnd()) {
+          task.status = TaskStatus.FAILED
+          fhirEngine.update(task)
+          task
+            .basedOn
+            .find { it.reference.startsWith(ResourceType.CarePlan.name) }
+            ?.extractId()
+            ?.takeIf { it.isNotBlank() }
+            ?.let {
+              val carePlan = fhirEngine.get<CarePlan>(it)
+              if (carePlan.isLastTask(task)) {
+                carePlan.status = CarePlan.CarePlanStatus.COMPLETED
+                fhirEngine.update(carePlan)
+              }
+            }
+        } else if (task.hasStarted() && task.status == TaskStatus.REQUESTED) {
+          task.status = TaskStatus.READY
+          fhirEngine.update(task)
+        }
+      }
+
+    Timber.i("Done task scheduling")
+  }
+
+  private fun CarePlan.isLastTask(task: Task) =
+    this.activity.last()?.outcomeReference?.last()?.extractId() == task.logicalId
 }
